implement modularity and review suggestions
